+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- jrl_release: add `--sign-tag` to sign the tag by @ahoarau
+
+
 ## [2.0.0] - 2026-06-29
 
 - Rework everything from scratch in v2/ opt-in directory by @ahoarau